Refactor Request Password react component
